CYA PT-4 for FT claim.
