+++ src/main/modules/draft-store/redisData.json
@@ -139,7 +139,7 @@
           }
         }
       ],
-      "totalClaimAmount": 1000,
+      "totalClaimAmount": 11000,
       "interestClaimFrom": "FROM_CLAIM_SUBMIT_DATE",
       "calculatedInterest": "| Description | Amount | \n |---|---| \n | Claim amount | £ 1000 | \n | Interest amount | £ 0.00 | \n | Total amount | £ 1000.00 |",
       "claimAmountBreakup": [
