+++ src/main/services/features/response/checkAnswers/hearingRequirementsSection/buildFastTrackHearingRequirements.ts
@@ -1,0 +1,72 @@
+import {Claim} from 'models/claim';
+import {SummarySection} from 'models/summaryList/summarySections';
+import {YesNo, YesNoUpperCamelCase} from 'form/models/yesNo';
+import {SummaryRow, summaryRow} from 'models/summaryList/summaryList';
+import {t} from 'i18next';
+import {constructResponseUrlWithIdParams} from 'common/utils/urlFormatter';
+import {
+  DQ_CONSIDER_CLAIMANT_DOCUMENTS_URL,
+  DQ_REQUEST_EXTRA_4WEEKS_URL,
+  DQ_TRIED_TO_SETTLE_CLAIM_URL,
+} from 'routes/urls';
+import {changeLabel} from 'common/utils/checkYourAnswer/changeButton';
+
+const getEmptyStringIfUndefined = (value: string): string => value || '';
+
+export const triedToSettleQuestion = (claim: Claim, claimId: string, lng: string): SummaryRow => {
+  const option = claim.directionQuestionnaire?.hearing?.triedToSettle?.option === YesNo.YES
+    ? YesNoUpperCamelCase.YES
+    : YesNoUpperCamelCase.NO;
+
+  return summaryRow(
+    t('PAGES.CHECK_YOUR_ANSWER.TRIED_TO_SETTLE', {lng}),
+    option,
+    constructResponseUrlWithIdParams(claimId, DQ_TRIED_TO_SETTLE_CLAIM_URL),
+    changeLabel(lng),
+  );
+};
+
+export const requestExtra4WeeksQuestion = (claim: Claim, claimId: string, lng: string): SummaryRow => {
+  const option = claim.directionQuestionnaire?.hearing?.requestExtra4weeks?.option === YesNo.YES
+    ? YesNoUpperCamelCase.YES
+    : YesNoUpperCamelCase.NO;
+
+  return summaryRow(
+    t('PAGES.CHECK_YOUR_ANSWER.REQUEST_EXTRA_4WEEKS', {lng}),
+    option,
+    constructResponseUrlWithIdParams(claimId, DQ_REQUEST_EXTRA_4WEEKS_URL),
+    changeLabel(lng),
+  );
+};
+
+export const considerClaimantDocQuestion = (claim: Claim, claimId: string, lng: string): SummaryRow => {
+  const option = claim.directionQuestionnaire?.hearing?.considerClaimantDocuments?.option === YesNo.YES
+    ? YesNoUpperCamelCase.YES
+    : YesNoUpperCamelCase.NO;
+
+  return summaryRow(
+    t('PAGES.CHECK_YOUR_ANSWER.CONSIDER_CLAIMANT_DOCUMENT', {lng}),
+    option,
+    constructResponseUrlWithIdParams(claimId, DQ_CONSIDER_CLAIMANT_DOCUMENTS_URL),
+    changeLabel(lng),
+  );
+};
+export const considerClaimantDocResponse = (claim: Claim, claimId: string, lng: string): SummaryRow => {
+  const details = claim.directionQuestionnaire?.hearing?.considerClaimantDocuments?.details;
+
+  return summaryRow(
+    t('PAGES.CHECK_YOUR_ANSWER.GIVE_DOC_DETAILS', {lng}),
+    getEmptyStringIfUndefined(details),
+  );
+};
+export const buildFastTrackHearingRequirements = (claim: Claim, hearingRequirementsSection: SummarySection, claimId: string, lng: string) => {
+
+  if (claim?.directionQuestionnaire?.hearing) {
+    hearingRequirementsSection.summaryList.rows.push(triedToSettleQuestion(claim, claimId, lng));
+    hearingRequirementsSection.summaryList.rows.push(requestExtra4WeeksQuestion(claim, claimId, lng));
+    hearingRequirementsSection.summaryList.rows.push(considerClaimantDocQuestion(claim, claimId, lng));
+
+    if(claim.directionQuestionnaire?.hearing?.considerClaimantDocuments.option == YesNo.YES)
+      hearingRequirementsSection.summaryList.rows.push(considerClaimantDocResponse(claim, claimId, lng));
+  }
+};
